Working on local pc

+++ cldm/model.py
@@ -25,4 +25,4 @@ def create_model(config_path):
     config = OmegaConf.load(config_path)
     model = instantiate_from_config(config.model).cpu()
     print(f'Loaded model config from [{config_path}]')
-    return model.half()
+    return model
